chunks: move _registerPrefix from cost,
we keep cost around for testing
